Bug 1334054 - fix CERT_FormatName output buffer length calculation r=franziskus
Summary:
Before this patch, CERT_FormatName attempted to account for the length of the
additional formatting in its output buffer length, but added an insufficient
amount (a fixed 128 bytes). This patch dynamically accounts for the additional
space required by the formatting output (it can over-account in some cases, but
this is unlikely to be a performance concern compared to the original
implementation).

Reviewers: franziskus

Differential Revision: https://nss-review.dev.mozaws.net/D307